+++ UCR.Plugins/Remapper/AxisToButton.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Reflection;
 using HidWizards.UCR.Core.Attributes;
 using HidWizards.UCR.Core.Models;
@@ -8,62 +9,80 @@ using HidWizards.UCR.Core.Utilities.AxisHelpers;
 
 namespace HidWizards.UCR.Plugins.Remapper
 {
-    [Plugin("Axis to Button", Group = "Button", Description = "Map from one axis to two buttons")]
+    [Plugin("Axis to Button", Group = "Button", Description = "Map from one pedal style axis to a button, with press and release threshold")]
     [PluginInput(DeviceBindingCategory.Range, "Axis")]
-    [PluginOutput(DeviceBindingCategory.Momentary, "Button high")]
-    [PluginOutput(DeviceBindingCategory.Momentary, "Button low")]
+    [PluginOutput(DeviceBindingCategory.Momentary, "Button")]
     public class AxisToButton : Plugin
     {
         [PluginGui("Invert", Order = 0)]
         public bool Invert { get; set; }
 
-        [PluginGui("Dead zone", Order = 1)]
-        public int DeadZone { get; set; }
+        [PluginGui("Press at axis value", Order = 1)]
+        public double PressPercent { get; set; }
 
-        private readonly DeadZoneHelper _deadZoneHelper = new DeadZoneHelper();
+        [PluginGui("Release at axis value", Order = 2)]
+        public double ReleasePercent { get; set; }
+
+        private bool _pressed;
+        private double _pressThresh;
+        private double _releaseThresh;
 
         public AxisToButton()
         {
-            DeadZone = 30;
+            PressPercent = -80;
+            ReleasePercent = -100;
         }
 
         public override void InitializeCacheValues()
         {
             Initialize();
+            _pressThresh = Functions.GetRangeFromPercentage(PressPercent);
+            _releaseThresh = Functions.GetRangeFromPercentage(ReleasePercent);
         }
 
         public override void Update(params short[] values)
         {
             var value = values[0];
             if (Invert) value = Functions.Invert(value);
-            switch (Math.Sign(_deadZoneHelper.ApplyRangeDeadZone(value)))
+            //Debug.WriteLine($"Current: {value}, Press @: {_pressThresh} Release @: {_releaseThresh}, Pressed: {_pressed}");
+            if (_pressed)
             {
-                case 0:
+                if (value <= _releaseThresh)
+                {
+                    _pressed = false;
                     WriteOutput(0, 0);
-                    WriteOutput(1, 0);
-                    break;
-                case 1:
+                }
+            }
+            else
+            {
+                if (value >= _pressThresh)
+                {
+                    _pressed = true;
                     WriteOutput(0, 1);
-                    WriteOutput(1, 0);
-                    break;
-                case -1:
-                    WriteOutput(0, 0);
-                    WriteOutput(1, 1);
-                    break;
+                }
             }
         }
 
         private void Initialize()
         {
-            _deadZoneHelper.Percentage = DeadZone;
         }
 
         public override PropertyValidationResult Validate(PropertyInfo propertyInfo, dynamic value)
         {
             switch (propertyInfo.Name)
             {
-                case nameof(DeadZone):
-                    return InputValidation.ValidatePercentage(value);
+                case nameof(PressPercent):
+                    if (value < ReleasePercent)
+                    {
+                        return new PropertyValidationResult(false, "Press must be higher or equal to Release");
+                    }
+                    return InputValidation.ValidateSignedPercentage(value);
+                case nameof(ReleasePercent):
+                    if (value > PressPercent)
+                    {
+                        return new PropertyValidationResult(false, "Release must be lower or equal to Press");
+                    }
+                    return InputValidation.ValidateSignedPercentage(value);
             }
 
             return PropertyValidationResult.ValidResult;